requirements.txt
certifi==2022.12.7
Flask==2.2.3
google_api_python_client==2.82.0
google_auth_oauthlib==1.0.0
pandas==1.5.3
pyairtable==1.4.0
python_dateutil==2.8.2
requests==2.28.2
setuptools==65.6.3
six==1.16.0
urllib3==1.26.14
gunicorn==20.1.0